bank-api/pyproject.toml
[tool.poetry]
name = "bank-api"
version = "0.1.0"
description = "API Bancária Assíncrona com FastAPI, JWT e SQLite"
authors = ["Seu Nome <jgsrocha@email.com>"]
readme = "README.md"
packages = [{ include = "bank_api", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"

fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlalchemy = "^2.0.29"
aiosqlite = "^0.20.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }

[tool.poetry.dev-dependencies]
# Adicione aqui ferramentas de linting/teste se quiser, ex:
# black = "^24.0.0"
# isort = "^5.13.0"
# pytest = "^8.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
